Use `usermod` instead of `lchsh` in custom.just
This fixes #408 (`just fish` and `just zsh` failing). The problem (and the reason that #420 doesn't fix it) is because the `lchsh` command actually doesn't accept the new shell as a CLI argument at all!

Another way to do this (and continue using `lchsh`) would be to instead run the commands as, e.g., `echo /usr/bin/zsh | sudo lchsh $USER`, but this results in an empty (and thus slightly misleading) prompt being shown to the user:

```console
$ echo /usr/bin/zsh | sudo lchsh $USER
[sudo] password for bri: 
Changing shell for bri.
New Shell [/usr/bin/zsh]: Shell changed.```

If we just use `usermod` it doesn't actually give any feedback, so I also threw in a little posix pipe to parse the shell back out of /etc/passwd after.