fix(-dx): Don't remove ganto-lxc4 repo.
Not all packages from ganto-lxc4 repo are installed on the image. Additionally, these missing packages conflict and can't be installed without changing the build script. lxd-tools and incus-tools are the packages in conflict.

This allows a user to layer the respective tools if necessary. Incus-tools provides tooling of both lxd-tools and lxd-migrate.