feat: bluefin-cli enable sudo/chown linuxbrew
Enables sudo in the bluefin-cli.  It also chowns the linuxbrew files to UID 1000. This makes the assumption that the user has UID 1000.

For whatever reason, the groups are not generated correctly for the user so we cannot reuse adm/wheel. The only valid group for the user is their own. Since most users are user 1000, we can assume this is the case. If they are not, they will have to use the su-exec method that already exists to change /etc/sudoers and chown linuxbrew.