feat(bluefin-cli): Persist State of Cellar
Persist the state of Homebrew's Cellar for #768

If the Cellar does not exist. It will be created in the User's home directory. It is then volume mounted into the container making it accessible on any distrobox enter for the container.

When using just bluefin-cli, it will relink all packages and reinstall explicitly installed packages.