Fix path
The extension was missing from the filepath, which led to the completions script running on every shell startup because the `brew` symlink never exists. 

That completion script takes nearly a second to complete, and I found this bug when I was trying to understand why fish took so long to startup.