chore(ci): always replace kernel packages
This simplifies the logic around installing cached kernels by always
doing it. This means that "vanilla"/"main" images will have the true
vanilla kernel replaced by essentially the same package, but the
repackaged version which we have signed in addition to the Fedora
signature.

This fixes a bug introduced in #1722 where accidentally stopped
installing the cached kernels for asus/surface images.