chore(ci): replace vanilla kernels not asus/surface
Fixes bug in #1752 which ignored the fact that asus/surface images in
Bluefin are sourced from the HWE images, thus they already have their
expected kernels.

I do think this end state is a bit nicer though as it is easier to grok
that this script only works on specific akmods flavors rather than the
nvidia_type which didn't seem to have a clear correlation to replacing
the kernel.