feat: add cryfs to bluefin
This makes io.github.mpobaschnig.Vaults functional. Although cryfs can
be installed by the user through brew or a distrobox, Vaults won't
detect it. It must be installed directly onto the system to be detected
by Vaults. Of the four months I've used Bluefin, this is the only
package I've been stuck with layering.

Why provide support for Vaults? KDE Plasma contains a built-in feature
called vaults that allows for the creation of locally encrypted
containers. This can offer a second layer of encryption beyond LUKS or
allow you to create encrypted containers on an otherwise unencrypted
file system. This feature was added in Plasma 5.11.

Vaults provides support for that feature on Gnome, helping bring feature
parity between the two desktop environments.

This commit only adds cryfs to Bluefin because it's already included in
KDE Plasma as a backend for its own vault feature. Yep, this means that
Aurora already has it! No need to add it again.