chore: Add new documentation URL
Note:  Aurora documentation is a work-in-progress, so this PR shouldn't be merged right away.

This is more of a proposal PR, but since Aurora has its own [documentation in development](https://docs.getaurora.dev/), then I propose a general 'landing page' that includes links to all 3 end-user images.  

This isn't the best solution since it is a desktop shortcut that requires the user to select which documentation they want based on whether they're using GNOME or KDE Plasma, so if there is a better method that someone can suggest then this proposal can be shot down.