fix: remove Alt-F5 from GNOME unmaximize keybinding
GNOME Settings only shows '<Super>Down' as the "unmaximize" keybinding
however, '<Alt>F5' is also there as a hidden default.

This is confusing enough, but the real trouble is this keybinding
conflicts with VSCode's default "next diff" shortcut.

As this hidden shortcut seems somewhat deprecated and conflicts with
VSCode, which we ship in DX, I'm removing it.