feat: update bluefin stable kernel pin to 6.12.9
Note: 6.12.9 is the current coreos-testing kernel and is known to have
fixed the AMD suspend bug.

We're switching the "bluefin:stable" builds to use this kernel since it
allows us to ship not-broken kernel and not-broken nvidia.

We don't have a pre-existing build for Fedora 40, so we can't ship a fix
for "bluefin:gts" at this time.