fix: pin mutter to 49.1 on f43
Seems that there was a regression on mutter that was being super
annoying.

From @adamisrael on discord:

> Has anyone noticed a change recently where when you
control-alt-<arrow>, the arrow key continues to repeat when you switch
workspace? Started happening to me yesterday, and continues today after
an update. It's quite annoying.

Please unpin whenever this is not necessary anymore. Someone should
periodically check this

Signed-off-by: Tulip Blossom <tulilirockz@outlook.com>