fix: cleanup /boot /tmp
bootc lint complains about /boot/symvers-*.xz which is part of the
kernel-core package.

There are leftovers of 600MiB in /tmp due to temporary build artifacts
like akmods, offline docs.

hhd/rechunk previously cleaned this up, so we never noticed.
This is necessary when we move off of this rechunker implementation.

The complaints about /var are expected due to the way we mount
/var/cache as type=cache.